Rename pricing.ts to cost-estimation.ts. Add a new test: test-cost-estimation.ts
